Adding real c/cpp testcases
