+++ .circleci/config.yml
@@ -11,14 +11,8 @@ jobs:
           command: |
             sudo sh -c 'echo "deb http://deb.debian.org/debian stretch-backports main" > /etc/apt/sources.list.d/backports.list'
             sudo apt-get update
-            sudo apt-get install -y cmake build-essential
+            sudo apt-get -t stretch-backports install cmake
       - run:
-          name: Build libra-dev
-          working_directory: libra-dev
+          name: Build everything
           command: |
-            cargo build
-      - run:
-          name: Build everything else
-          command: |
-            mkdir build
-            cd build
+            ./build.sh
