+++ .gitignore
@@ -3,9 +3,11 @@
 
 # cmake stuff
 cmake-build-*
+/build
 
 # Rust specific ignores
 # Please follow https://help.github.com/en/articles/ignoring-files to create a global
 # .gitignore file locally for IDE/Emacs/Vim generated files.
 **/target
 **/*.rs.bk
+
