+++ README.md
@@ -1,2 +1,5 @@
 # libra-client-dev
 Core rust client library for Libra
+
+# How do I even
+./build.sh
